Actualizar el viaje terminado

+++ src/viajes/Viaje.java
@@ -20,7 +20,7 @@ public class Viaje {
 	/**
 	 * Atributo para guardar el precio del viaje
 	 */
-	int precio = 0;
+	double precio = 0;
 
 	/**
 	 * Constructor para crear objetos con todos sus atributos
@@ -29,7 +29,7 @@ public class Viaje {
 	 * @param fecha  La fecha del viaje
 	 * @param precio Precio del viaje
 	 */
-	public Viaje(String lugar, String fecha, int precio) {
+	public Viaje(String lugar, String fecha, double precio) {
 		if (lugar != null && !lugar.equals("")) {
 			this.lugar = lugar;
 		}
@@ -116,7 +116,7 @@ public class Viaje {
 				Integer.parseInt(fecha.substring(3, 5));
 			} // Fin Case 2
 			
-			// Case 3: analiza el aÃ±o
+			// Case 3: analiza el aÃÂ±o
 			case 3 -> {
 				Integer.parseInt(fecha.substring(6, 10));
 			} // Fin Case 3
@@ -196,7 +196,7 @@ public class Viaje {
 	 * 
 	 * @return Devuelve el precio del viaje
 	 */
-	public int getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
 
@@ -205,12 +205,12 @@ public class Viaje {
 	 * 
 	 * @param precio Precio con el que actualizar el viaje
 	 */
-	public void setPrecio(int precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
 	/**
-	 * Función para mostrar la información de cada viaje
-	 * @return Devuelve una cadena con la información de los viajes
+	 * FunciÃ³n para mostrar la informaciÃ³n de cada viaje
+	 * @return Devuelve una cadena con la informaciÃ³n de los viajes
 	 */
 	@Override
 	public String toString() {
@@ -226,8 +226,8 @@ public class Viaje {
 	}
 	
 	/**
-	 * Función para comnparar dos viajes y saber si son iguales
-	 * @param o Recibe un objeto que se casteará a uno de tipo viaje para poder comparar dos viajes
+	 * FunciÃ³n para comnparar dos viajes y saber si son iguales
+	 * @param o Recibe un objeto que se castearÃ¡ a uno de tipo viaje para poder comparar dos viajes
 	 * @return Devuelve true si la fecha y lugar son iguales y false en caso contrario
 	 */
 	@Override
